% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{pl_login}
\alias{pl_login}
\title{Authenticate with Plone}
\usage{
pl_login(plone)
}
\arguments{
\item{plone}{Location of the Plone server}
}
\description{
Adds a JSON Web Token to the session environment variables
}
