% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{pl_login_renew}
\alias{pl_login_renew}
\title{Renews the JSON Web Token for a previously-authenticated user}
\usage{
pl_login_renew(plone)
}
\description{
Renews the JSON Web Token for a previously-authenticated user
}
